+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## September 9, 2020
+- Modern JavaScript From The Beginning (Udemy)
+  - Section 7: Asynchronous JavaScript, Ajax & Fetch API
+    - Callback Functions
+- An example with a callback function was added.
+
 ## September 8, 2020
 - Modern JavaScript From The Beginning (Udemy)
   - Section 7: Asynchronous JavaScript, Ajax & Fetch API